Got rid of np.float to make numpy happy

+++ src/skmultiflow/utils/utils.py
@@ -153,7 +153,7 @@ def is_scalar_nan(x):
     """Tests if x is NaN
 
     This function is meant to overcome the issue that np.isnan does not allow
-    non-numerical types as input, and that np.nan is not np.float('nan').
+    non-numerical types as input, and that np.nan is not np.float64('nan').
 
     Parameters
     ----------
